fix(change-risk): credit the crate's own plural _tests.rs convention
Nine existing flat-module test files (artifacts_tests.rs,
audit_report/model_tests.rs, ...) use the `_tests.rs` plural suffix;
zero use the singular `_test.rs` is_test_file actually checked. Every
PR adding one of these -- including this session's own
boundary_rules_tests.rs -- scored "no tests touched" and got flagged
as a change-risk violation despite genuine coverage. Same class of bug
as a3ac9a1 (bare tests.rs), just the plural sibling never covered.

Split change_risk/tests.rs into tests/{mod,predicates,scoring}.rs in
the same pass: the new test pushed it over the god-file ratchet
(26 functions, 464 loc) the moment it landed on a tree that hasn't
merged #202's identical split yet.
